pyproject.toml
[project]
name = "LeetSquad"
version = "0.1.0"
description = "Green (eval) agent that benchmarks white (participant) agents on code generation capabilities using Leetcode problems."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "a2a_sdk[http-server]>=0.3.10",
    "boto3>=1.40.0",
    "click>=8.1.0",
    "fastapi>=0.100.0",
    "ruff>=0.14.0",
    "uvicorn>=0.23.0",
    "earthshaker",
    "dotenv",
    "python-dotenv>=1.2.1",
    "openai>=2.8.0",
    "openai-agents>=0.6.1",
]
